chore: split previewer
