+++ .dumi/theme/builtins/Previewer/index.tsx
@@ -1,21 +1,10 @@
+import React, { Suspense } from 'react';
 import type { IPreviewerProps } from 'dumi';
-import { useTabMeta } from 'dumi';
-import React from 'react';
-import CodePreviewer from './CodePreviewer';
-import DesignPreviewer from './DesignPreviewer';
 
-export interface AntdPreviewerProps extends IPreviewerProps {
-  originDebug?: IPreviewerProps['debug'];
-}
+const Previewer = React.lazy(() => import('./Previewer'));
 
-const Previewer: React.FC<AntdPreviewerProps> = (props) => {
-  const tab = useTabMeta();
-
-  if (tab?.frontmatter.title === 'Design') {
-    return <DesignPreviewer {...props} />;
-  }
-
-  return <CodePreviewer {...props} />;
-};
-
-export default Previewer;
+export default (props: IPreviewerProps) => (
+  <Suspense fallback={null}>
+    <Previewer {...props} />
+  </Suspense>
+);
